"Fixed variable scoping and removed unnecessary data from response"
The code diff shows that the variable user has been moved outside of the get request callback function, ensuring that its value is persisted between requests. Additionally, the data property has been removed from the response object, as it was not defined in the surrounding code. The commit also removes the process.env.DATA\_BASE\_URL from the response object.