"Update authentication and add sign in token verification for admin routes"
This diff includes updates to the admin controllers, models, types, and routes for authentication and token verification. The `SignUp` and `SignIn` functions in the `admin` controller have been updated to use a new `generateAuthToken` function from the `JwtToken` utility. The `admin` model has been updated to include a `signInToken` field instead of a `loginToken`. A new `checkSignInToken` function has been added to the `admin` controller to verify a given sign in token. The `index` file in the `routes/admin` directory has been updated to include a new route for token verification. Lastly, the `ApiResponseProps` type in the `types` directory has been updated to include a new `signInToken` field. A new `updateSignInToken` function has been added to the `helper` utility to update the sign in token for a user in the database.