"Add quiz object to routes data with CRUD operations for quiz and question"
This commit adds a new object "quiz" to the "data" array in the "routs.ts" file. The "quiz" object contains a "preUrl" property set to "quiz/", an empty "solved" object, and an array of "unsolved" routes for CRUD operations related to quizzes and questions. The "unsolved" array includes routes for creating a quiz, creating a question, getting all quizzes, getting all admin quizzes, updating a quiz, and updating a question.